requirements.txt
boto
gevent
pyyaml

